fix bug

+++ ft_nbrlen.c
@@ -6,7 +6,7 @@
 /*   By: joesanto <joesanto@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/06 17:22:41 by joesanto          #+#    #+#             */
-/*   Updated: 2025/10/08 11:41:55 by joesanto         ###   ########.fr       */
+/*   Updated: 2025/10/08 11:43:33 by joesanto         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_nbrlen(ssize_t nbr, int base)
 {
-	int	len;
+	size_t	len;
 
 	if (base < 2)
 		return (0);
